+++ src/providers/loading-ionic.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { PluginBase, PostRequestErrorPlugin, PostRequestPlugin, PreRequestPlugin } from '@ramonornela/http';
+import { PluginBase, PostRequestErrorPlugin, PostRequestPlugin, PreRequestPlugin } from '@mbamobi/http';
 import { Loading, LoadingController } from 'ionic-angular';
 
 @Injectable()
@@ -53,7 +53,7 @@ export class LoadingIonicPlugin extends PluginBase implements PreRequestPlugin, 
 
   dismiss() {
     if (this.loading) {
-      this.loading.dismissAll();
+      this.loading.dismiss();
     }
 
     this.loading = null;